Make sure 3-argument pow() is always called with ints
Python does not allow overriding 3-arg pow() when the first argument is an int
(see https://docs.python.org/3/reference/datamodel.html#object.__rpow__), so
if the second or thired argument can be a SymPy Integer, we need to make sure
they are converted to int first.

+++ sympy/ntheory/residue_ntheory.py
@@ -654,8 +654,8 @@ def _is_nthpow_residue_bign(a, n, m):
                 return False
         return True
     f = totient(m)
-    k = f // igcd(f, n)
-    return pow(a, k, m) == 1
+    k = int(f // igcd(f, n))
+    return pow(a, k, int(m)) == 1
 
 
 def _is_nthpow_residue_bign_prime_power(a, n, p, k):
